ugps: add quality measurement parameters
The current version of ugps does not report any quality parameters for GPS receiver fixation. Considering the positioning error, which can be up to hundreds of meters, it is helpful to know how accurately the GPS determines the position.
This PR supports determining the number of satellites and the HDOP (Horizontal Dilution of Precision) parameter to address this.
The GPGGA sentence is already parsed in the code, so it just extends the amount of data it reads.

Fixes: #2 (ugps miss GPS quality parameters)
Signed-off-by: Zbynek Kocur <zbynek.kocur@fel.cvut.cz>